+++ src/ace_time/LocalDateTime.h
@@ -26,6 +26,7 @@ namespace ace_time {
  * Parts of this class were inspired by the java.time.LocalDateTime class of
  * Java 11
  * (https://docs.oracle.com/en/java/javase/11/docs/api/java.base/java/time/LocalDateTime.html).
+ * The 'fold' parameter was inspired by the datetime package in Python 3.6.
  */
 class LocalDateTime {
   public:
@@ -39,21 +40,25 @@ class LocalDateTime {
      * @param hour hour [0-23]
      * @param minute minute [0-59]
      * @param second second [0-59], does not support leap seconds
+     * @param fold optional disambiguation of multiple occurences [0, 1]
      */
     static LocalDateTime forComponents(int16_t year, uint8_t month,
-        uint8_t day, uint8_t hour, uint8_t minute, uint8_t second) {
+        uint8_t day, uint8_t hour, uint8_t minute, uint8_t second,
+        uint8_t fold = 0) {
       int8_t yearTiny = LocalDate::isYearValid(year)
           ? year - LocalDate::kEpochYear
           : LocalDate::kInvalidYearTiny;
-      return forTinyComponents(yearTiny, month, day, hour, minute, second);
+      return forTinyComponents(
+          yearTiny, month, day, hour, minute, second, fold);
     }
 
     /** Factory method using components with an int8_t yearTiny. */
     static LocalDateTime forTinyComponents(int8_t yearTiny, uint8_t month,
-        uint8_t day, uint8_t hour, uint8_t minute, uint8_t second) {
+        uint8_t day, uint8_t hour, uint8_t minute, uint8_t second,
+        uint8_t fold = 0) {
       return LocalDateTime(
           LocalDate::forTinyComponents(yearTiny, month, day),
-          LocalTime::forComponents(hour, minute, second));
+          LocalTime::forComponents(hour, minute, second, fold));
     }
 
     /**
@@ -67,7 +72,9 @@ class LocalDateTime {
      *    (2000-01-01 00:00:00). Use LocalDate::kInvalidEpochSeconds to define
      *    an invalid instance whose isError() returns true.
      */
-    static LocalDateTime forEpochSeconds(acetime_t epochSeconds) {
+    static LocalDateTime forEpochSeconds(
+        acetime_t epochSeconds, uint8_t fold = 0) {
+
       LocalDate ld;
       LocalTime lt;
       if (epochSeconds == LocalDate::kInvalidEpochSeconds) {
@@ -83,7 +90,7 @@ class LocalDateTime {
         // epochSeconds could be negative.
         acetime_t seconds = epochSeconds - 86400 * days;
         ld = LocalDate::forEpochDays(days);
-        lt = LocalTime::forSeconds(seconds);
+        lt = LocalTime::forSeconds(seconds, fold);
       }
 
       return LocalDateTime(ld, lt);
@@ -213,6 +220,12 @@ class LocalDateTime {
 
     /** Set the second. */
     void second(uint8_t second) { mLocalTime.second(second); }
+
+    /** Return the fold. */
+    uint8_t fold() const { return mLocalTime.fold(); }
+
+    /** Set the fold. */
+    void fold(uint8_t fold) { mLocalTime.fold(fold); }
 
     /** Return the day of the week, Monday=1, Sunday=7 (per ISO 8601). */
     uint8_t dayOfWeek() const { return mLocalDate.dayOfWeek(); }